requirements.txt
cookiecutter
click==3.3
toml
jinja2
python-slugify
markdown
